(refactor) Change project structure
