+++ src/frontend/packages/core/src/components/livekit/assets/icons.tsx
