+++ src/frontend/packages/core/src/components/livekit/controls/ControlBar/ControlBar.tsx
@@ -1,16 +1,15 @@
-import { CameraDisabledIcon, CameraIcon, ChatIcon, LeaveIcon, MediaDeviceMenu, MicDisabledIcon, MicIcon, ScreenShareIcon, ScreenShareStopIcon, TrackToggleProps, UseChatToggleProps, useChatToggle, useDisconnectButton, useLayoutContext, useLocalParticipantPermissions, useRemoteParticipants, useTrackToggle } from "@livekit/components-react"
+import { CameraDisabledIcon, CameraIcon, ChatIcon, LeaveIcon, MediaDeviceMenu, MicDisabledIcon, MicIcon, ScreenShareIcon, ScreenShareStopIcon, TrackToggleProps, UseChatToggleProps, useChatToggle, useDisconnectButton, useLocalParticipantPermissions, useRemoteParticipants, useTrackToggle } from "@livekit/components-react"
 import { Button, ToastProps, VariantType, defaultTokens } from "@openfun/cunningham-react"
 import { Track } from "livekit-client"
 import { Card, DropButton } from "grommet"
-import { MouseEventHandler, useEffect, useState } from "react"
-import { ParticipantLayoutToggle, RaiseHand, useParticipantLayoutContext } from "./participants"
-import { useAudioAllowed, useScreenSharingAllowed, useVideoAllowed } from "../utils/hooks"
-import { useIsMobile, useIsSmallSize } from "../../../hooks/useIsMobile"
-import { Event, useEventHandler } from "../../../services/livekit/events"
-import { LayoutToggle } from "../conference/conference"
-import { MoreIcon } from "../utils/icons"
-import { tokens } from "../../../cunningham-tokens"
-import { useMagnifyRoomContext } from "../../../context/room"
+import { MouseEventHandler } from "react"
+import { useIsMobile, useIsSmallSize } from "../../../../hooks/useIsMobile"
+import { Event, useEventHandler } from "../../../../services/livekit/events"
+import { MoreIcon } from "../../assets/icons"
+import { tokens } from "../../../../cunningham-tokens"
+import { LayoutToggle } from "../actions/togglers/LayoutToggle"
+import { HandRaiseToggle } from "../actions/togglers/HandRaiseToggle"
+import { ParticipantLayoutToggle } from "../actions/togglers/ParticipantLayoutToggle"
 
 
 
@@ -63,15 +62,6 @@ const defaultControlBarProps: ControlBarProps = {
     audioControl: true,
     screenSharingControl: true
 }
-
-export const MagnifyControlBar = () => {
-    const permissions = useLocalParticipantPermissions()
-    const video = useVideoAllowed(permissions)
-    const audio = useAudioAllowed(permissions)
-    const screenSharing = useScreenSharingAllowed(permissions)
-    return <ControlBar videoControl={video} audioControl={audio} screenSharingControl={screenSharing} />
-}
-
 
 export const ControlBar = (props: ControlBarProps) => {
 
@@ -128,7 +118,7 @@ export const ControlBar = (props: ControlBarProps) => {
     const mobileSelector =
         <div style={{ color: "white", backgroundColor: `${tokens.theme.colors["primary-400"]}`, flexDirection: "row", display: "flex", alignItems: "center", justifyContent: "center" }}>
             <LayoutToggle />
-            <RaiseHand />
+            <HandRaiseToggle />
         </div>
 
     return (
@@ -136,7 +126,7 @@ export const ControlBar = (props: ControlBarProps) => {
             {!mobile &&
                 <>
                     <Card style={{ borderRadius: "0.6em", display: "flex", flexDirection: "row" }} className="bg-primary-400">
-                        <RaiseHand />
+                        <HandRaiseToggle />
                     </Card>
                     <Card style={{ borderRadius: "0.6em", display: "flex", flexDirection: "row" }} className="bg-primary-400">
                         <ParticipantLayoutToggle />
