+++ src/frontend/packages/core/src/components/livekit/utils/permissions.tsx
@@ -1,5 +1,3 @@
-import { useLocalParticipant } from "@livekit/components-react"
-import { Participant } from "livekit-client"
 import { ParticipantPermission } from "livekit-client/dist/src/proto/livekit_models_pb"
 
 export const useVideoAllowed = (permissions : ParticipantPermission | undefined) : boolean => {
@@ -13,3 +11,4 @@ export const useAudioAllowed = (permissions : ParticipantPermission | undefined)
 export const useScreenSharingAllowed = (permissions : ParticipantPermission | undefined) : boolean => {
     return (permissions?.canPublish && (permissions?.canPublishSources ?? [3, 4]).includes(3) && (permissions?.canPublishSources ?? [3, 4]).includes(4)) ?? true
 }
+
